+++ actix-rt/CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `actix_rt::ArbiterBuilder` to allow user to configure the thread spawned for the arbiter.
+- Add `Arbiter::alive` and `ArbiterHandle::alive` to check is the arbiter is still alive.
 
 ## 2.10.0